requirements.txt
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
defer==1.0.6
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
geographiclib==1.49
geopy==1.20.0
httplib2==0.9.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
numpy==1.16.4
onboard==1.4.1
PAM==0.4.2
pandas==0.25.0
pep8==1.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyodbc==4.0.26
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.3
python-xapp==1.4.0
python-xlib==0.20
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.3.6
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.22
Werkzeug==0.15.5
xkit==0.0.0
